Undo changes and restore to original state.
- Misunderstood the issue at first which caused incorrect DB changes.
- Reverted them and kept only the correct ones.
- sorry for that (:

+++ openlibrary/core/infobase_schema.sql
@@ -34,7 +34,7 @@ create index thing_created_idx ON thing(created);
 create unique index thing_key_idx ON thing(key);
 
 create table transaction (
-    id bigserial primary key,
+    id serial primary key,
     action varchar(256),
     author_id bigint references thing,
     ip inet,
@@ -52,7 +52,7 @@ create index transaction_ip_idx ON transaction(ip);
 create index transaction_created_idx ON transaction(created);
 
 create table transaction_index (
-    tx_id bigint references transaction,
+    tx_id int references transaction,
     key text,
     value text
 );
@@ -61,21 +61,21 @@ create index transaction_index_key_value_idx ON transaction_index(key, value);
 create index transaction_index_tx_id_idx ON transaction_index(tx_id);
 
 create table version (
-    id bigserial primary key,
+    id serial primary key,
     thing_id bigint references thing,
     revision int,
-    transaction_id bigint references transaction,
+    transaction_id int references transaction,
     UNIQUE (thing_id, revision)
 );
 
 create table property (
-    id bigserial primary key,
+    id serial primary key,
     type bigint references thing,
     name text,
     UNIQUE (type, name)
 );
 
-CREATE FUNCTION get_property_name(bigint, bigint)
+CREATE FUNCTION get_property_name(bigint, int)
 RETURNS text AS
 'select property.name FROM property, thing WHERE thing.type = property.type AND thing.id=$1 AND property.id=$2;'
 LANGUAGE SQL;
@@ -106,7 +106,7 @@ create unique index data_thing_id_revision_idx ON data(thing_id, revision);
 
 create table author_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -115,7 +115,7 @@ create index author_boolean_thing_id_idx ON author_boolean(thing_id);
 
 create table author_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -124,7 +124,7 @@ create index author_int_thing_id_idx ON author_int(thing_id);
 
 create table author_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -133,7 +133,7 @@ create index author_ref_thing_id_idx ON author_ref(thing_id);
 
 create table author_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -142,7 +142,7 @@ create index author_str_thing_id_idx ON author_str(thing_id);
 
 create table datum_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -151,7 +151,7 @@ create index datum_int_thing_id_idx ON datum_int(thing_id);
 
 create table datum_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -160,7 +160,7 @@ create index datum_ref_thing_id_idx ON datum_ref(thing_id);
 
 create table datum_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -169,7 +169,7 @@ create index datum_str_thing_id_idx ON datum_str(thing_id);
 
 create table edition_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -178,7 +178,7 @@ create index edition_boolean_thing_id_idx ON edition_boolean(thing_id);
 
 create table edition_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -187,7 +187,7 @@ create index edition_int_thing_id_idx ON edition_int(thing_id);
 
 create table edition_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -196,7 +196,7 @@ create index edition_ref_thing_id_idx ON edition_ref(thing_id);
 
 create table edition_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -205,7 +205,7 @@ create index edition_str_thing_id_idx ON edition_str(thing_id);
 
 create table publisher_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -214,7 +214,7 @@ create index publisher_boolean_thing_id_idx ON publisher_boolean(thing_id);
 
 create table publisher_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -223,7 +223,7 @@ create index publisher_int_thing_id_idx ON publisher_int(thing_id);
 
 create table publisher_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -232,7 +232,7 @@ create index publisher_ref_thing_id_idx ON publisher_ref(thing_id);
 
 create table publisher_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -241,7 +241,7 @@ create index publisher_str_thing_id_idx ON publisher_str(thing_id);
 
 create table scan_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -250,7 +250,7 @@ create index scan_boolean_thing_id_idx ON scan_boolean(thing_id);
 
 create table scan_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -259,7 +259,7 @@ create index scan_int_thing_id_idx ON scan_int(thing_id);
 
 create table scan_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -268,7 +268,7 @@ create index scan_ref_thing_id_idx ON scan_ref(thing_id);
 
 create table scan_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -277,7 +277,7 @@ create index scan_str_thing_id_idx ON scan_str(thing_id);
 
 create table subject_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -286,7 +286,7 @@ create index subject_boolean_thing_id_idx ON subject_boolean(thing_id);
 
 create table subject_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -295,7 +295,7 @@ create index subject_int_thing_id_idx ON subject_int(thing_id);
 
 create table subject_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -304,7 +304,7 @@ create index subject_ref_thing_id_idx ON subject_ref(thing_id);
 
 create table subject_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -312,7 +312,7 @@ create index subject_str_idx ON subject_str(key_id, value);
 create index subject_str_thing_id_idx ON subject_str(thing_id);
 
 create table type_int (
-    thing_id int references thing,
+    thing_id bigint references thing,
     key_id int references property,
     value int,
     ordering int default NULL
@@ -322,7 +322,7 @@ create index type_int_thing_id_idx ON type_int(thing_id);
 
 create table type_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -331,7 +331,7 @@ create index type_ref_thing_id_idx ON type_ref(thing_id);
 
 create table type_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -340,7 +340,7 @@ create index type_str_thing_id_idx ON type_str(thing_id);
 
 create table user_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -349,7 +349,7 @@ create index user_int_thing_id_idx ON user_int(thing_id);
 
 create table user_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -358,7 +358,7 @@ create index user_ref_thing_id_idx ON user_ref(thing_id);
 
 create table user_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -367,7 +367,7 @@ create index user_str_thing_id_idx ON user_str(thing_id);
 
 create table work_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -376,7 +376,7 @@ create index work_boolean_thing_id_idx ON work_boolean(thing_id);
 
 create table work_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -385,7 +385,7 @@ create index work_int_thing_id_idx ON work_int(thing_id);
 
 create table work_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -394,7 +394,7 @@ create index work_ref_thing_id_idx ON work_ref(thing_id);
 
 create table work_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -403,7 +403,7 @@ create index work_str_thing_id_idx ON work_str(thing_id);
 
 create table tag_boolean (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value boolean,
     ordering int default NULL
 );
@@ -412,7 +412,7 @@ create index tag_boolean_thing_id_idx ON tag_boolean(thing_id);
 
 create table tag_int (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value int,
     ordering int default NULL
 );
@@ -421,7 +421,7 @@ create index tag_int_thing_id_idx ON tag_int(thing_id);
 
 create table tag_ref (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value bigint references thing,
     ordering int default NULL
 );
@@ -430,7 +430,7 @@ create index tag_ref_thing_id_idx ON tag_ref(thing_id);
 
 create table tag_str (
     thing_id bigint references thing,
-    key_id bigint references property,
+    key_id int references property,
     value varchar(2048),
     ordering int default NULL
 );
@@ -455,7 +455,7 @@ create table store (
 );
 
 create table store_index (
-    id bigserial primary key,
+    id serial primary key,
     store_id bigint references store,
     type text,
     name text,
@@ -466,7 +466,7 @@ create index store_index_store_id_idx ON store_index (store_id);
 create index store_idx ON store_index(type, name, value);
 
 create table seq (
-    id bigserial primary key,
+    id serial primary key,
     name text unique,
     value int default 0
 );
@@ -482,7 +482,7 @@ $$ LANGUAGE SQL IMMUTABLE;
 CREATE INDEX thing_olid_idx ON thing(get_olid(key));
 
 CREATE TABLE stats (
-    id bigserial primary key,
+    id serial primary key,
     key text unique,
     type text,
     created timestamp without time zone,
@@ -494,7 +494,7 @@ CREATE INDEX stats_created_idx ON stats(created);
 CREATE INDEX stats_updated_idx ON stats(updated);
 
 CREATE TABLE waitingloan (
-    id bigserial primary key,
+    id serial primary key,
     book_key text,
     user_key text,
     status text default 'waiting',
@@ -512,7 +512,7 @@ CREATE INDEX waitingloan_status_idx ON waitingloan(status);
 
 
 CREATE TABLE import_batch (
-    id bigserial primary key,
+    id serial primary key,
     name text,
     submitter text,
     submit_time timestamp without time zone default (current_timestamp at time zone 'utc')
@@ -523,8 +523,8 @@ CREATE INDEX import_batch_submitter_idx ON import_batch(submitter);
 CREATE INDEX import_batch_submit_time_idx ON import_batch(submit_time);
 
 CREATE TABLE import_item (
-    id bigserial primary key,
-    batch_id bigint references import_batch,
+    id serial primary key,
+    batch_id int references import_batch,
     added_time timestamp without time zone default (current_timestamp at time zone 'utc'),
     import_time timestamp without time zone,
     status text default 'pending',
